backend/src/requirements.txt
Flask
gunicorn
google-cloud-firestore
flask-cors
requests
python-dotenv
